refactor: combine index and services exports of client
Otherwise when shell bundled the AgentHostingProvider could not find the ClientContext
