+++ packages/apps/plugins/plugin-help/src/components/HelpContextProvider/HelpContextProvider.tsx
@@ -6,8 +6,7 @@ import React, { type PropsWithChildren, useState, useEffect } from 'react';
 import Joyride, { ACTIONS, EVENTS } from 'react-joyride';
 
 import { usePlugins, resolvePlugin, parseLayoutPlugin } from '@dxos/app-framework';
-import { useShellDisplay } from '@dxos/react-client';
-import { ShellDisplay } from '@dxos/react-client/services';
+import { useShellDisplay, ShellDisplay } from '@dxos/react-client';
 // import { useThemeContext } from '@dxos/react-ui';
 import { tailwindConfig, type TailwindConfig } from '@dxos/react-ui-theme';
