+++ packages/devtools/cli-base/src/base.ts
@@ -13,9 +13,8 @@ import readline from 'node:readline';
 import pkgUp from 'pkg-up';
 
 import { type Daemon, LaunchctlRunner, PhoenixDaemon, SystemctlRunner, SystemDaemon } from '@dxos/agent';
-import { Client, Config } from '@dxos/client';
+import { Client, Config, fromAgent } from '@dxos/client';
 import { type Space } from '@dxos/client/echo';
-import { fromAgent } from '@dxos/client/services';
 import {
   DX_CONFIG,
   DX_DATA,
