+++ packages/sdk/react-client/package.json
@@ -49,13 +49,6 @@
       "node": "./dist/lib/node/mesh/index.cjs",
       "types": "./dist/types/src/mesh/index.d.ts"
     },
-    "./services": {
-      "browser": "./dist/lib/browser/services/index.mjs",
-      "import": "./dist/lib/browser/services/index.mjs",
-      "require": "./dist/lib/node/services/index.cjs",
-      "node": "./dist/lib/node/services/index.cjs",
-      "types": "./dist/types/src/services/index.d.ts"
-    },
     "./testing": {
       "browser": "./dist/lib/browser/testing/index.mjs",
       "import": "./dist/lib/browser/testing/index.mjs",
@@ -88,9 +81,6 @@
       ],
       "mesh": [
         "dist/types/src/mesh/index.d.ts"
-      ],
-      "services": [
-        "dist/types/src/services/index.d.ts"
       ],
       "testing": [
         "dist/types/src/testing/index.d.ts"
