+++ packages/sdk/react-client/project.json
@@ -17,7 +17,6 @@
           "packages/sdk/react-client/src/index.ts",
           "packages/sdk/react-client/src/invitations/index.ts",
           "packages/sdk/react-client/src/mesh/index.ts",
-          "packages/sdk/react-client/src/services/index.ts",
           "packages/sdk/react-client/src/testing/index.ts",
           "packages/sdk/react-client/src/worker.ts"
         ]
