+++ packages/sdk/react-client/src/client/AgentHostingProvider.tsx
@@ -4,8 +4,8 @@
 
 import React, { createContext, type PropsWithChildren, useContext, useState } from 'react';
 
+import { type AgentHostingProviderClient, AgentManagerClient, FakeAgentHostingProvider } from '@dxos/client';
 import { type Halo } from '@dxos/client/halo';
-import { type AgentHostingProviderClient, AgentManagerClient, FakeAgentHostingProvider } from '@dxos/client/services';
 import { type Config } from '@dxos/config';
 import { log } from '@dxos/log';
