+++ packages/sdk/react-client/src/client/index.ts
@@ -2,6 +2,7 @@
 // Copyright 2020 DXOS.org
 //
 
+export * from './AgentHostingProvider';
 export * from './ClientContext';
 export * from './useClientServices';
 export * from './useConfig';
