--- packages/sdk/react-client/src/services/index.ts
@@ -1,7 +1,0 @@
-//
-// Copyright 2024 DXOS.org
-//
-
-export * from '@dxos/client/services';
-
-export * from './AgentHostingProvider';
