+++ packages/sdk/shell/src/composites/Shell/Shell.tsx
@@ -6,14 +6,14 @@ import React, { useEffect, useState } from 'react';
 
 import { log } from '@dxos/log';
 import { useClient } from '@dxos/react-client';
-import { useSpace } from '@dxos/react-client/echo';
 import {
-  type InvitationUrlRequest,
-  type LayoutRequest,
   ShellDisplay,
   ShellLayout,
+  type InvitationUrlRequest,
+  type LayoutRequest,
   type ShellRuntime,
-} from '@dxos/react-client/services';
+} from '@dxos/react-client';
+import { useSpace } from '@dxos/react-client/echo';
 
 import { IdentityDialog } from '../IdentityDialog';
 import { JoinDialog } from '../JoinDialog';
