+++ packages/sdk/shell/src/composites/Shell/ShellContext.tsx
@@ -13,10 +13,9 @@ import React, {
   useState,
 } from 'react';
 
-import { type PublicKey, useClient } from '@dxos/react-client';
+import { type PublicKey, useClient, ShellDisplay, ShellLayout, type LayoutRequest } from '@dxos/react-client';
 import type { Space } from '@dxos/react-client/echo';
 import { useIdentity } from '@dxos/react-client/halo';
-import { ShellDisplay, ShellLayout, type LayoutRequest } from '@dxos/react-client/services';
 import { mx } from '@dxos/react-ui-theme';
 
 import { Shell } from './Shell';
